Fixes up group_relcast to have proper reliable behavior.
This is a lot of change but it ends up making reliable broadcast work
for the common usecase. There _are_ duplicate messages sine we can't
quite address the 2 generals problem with group workers connecting to
eachother in parallel. Yet.

Also adds a function to dump the message queue for a relcast server to
aid in debugging group issues.