Allow nat type to be set multiple times
The update_nat_type guard was only allowing the nat type to be changed
if it was different from the current known state. The update_this_peer
code has improved enough to nto update the peer if it hasn't changed,
so we can drop the guard and just update the nat type at will.