Extremely speculative attempt to handle simultaneous TCP connections
If 2 peers try to dial each other with the local/remote addresses being
the same, which happens because of reuseport, the typical TCP accept can
be bypassed and you get 2 clients connected to each other. This
complicates the session protocols like yamux, which want to know who is
the 'server'. This patch attempts to determine a 'server' when a
simultaneous connection happens by using random timeouts. If one client
has to send to make the handshake happen then it determines itself the
server and the connection is initialized via the multistream_server
path.