Detect fast closing open streams
When streams get assigned and then close really quickly we can end up
in a quick connecting/connected cycle which escalates pretty quickly
on the network. We try to slow this down by not canceling the connect
retry backoff until the worker has been in the connected state for
some period of time.