Attempt to only ACK relcast packets when necessary
Attempt to use 'range acks' in relcast whenever possible. If we get a
data packet flagged 'last' from the sender, and we're able to deliver
it, we do a 'range' ack on that packet (but not any prior packets).

If we have a canary packet (defer queue full) we ack the last un-acked
packet for that worker and then single-ack all subsequent deliverable
packets until we're able to deliver the canary.