Fix a socket/process leak in proxy
Before this change, we left a gen_server hanging around on the proxy server.
Unfortunately, this was the process that ranch thought owned the socket.
So, once the socket closed, the gen_server remained alive and so the
number of sockets ranch thought were open never decremented. This change
ensures that the socket owners are cleaned up and a test is extended to
verify this.