Fixes for cases where swarm has no initial connectivity
When the swarm comes up on a machine with no IPs, it can get in a state
where it takes a very long time (5 minutes) to detect an internet
connection because of the gossip targeting backoffs. This patch tweaks
the retargeting behaviour for seed connections and additionally will run
NAT discovery if an IP address change is detected; including going from
no IPs to any IPs.