Store peerbook address keys reversed for better random distribution
Storing peer address keys with the type bytes at the start leads to
extremely bad distribution of the keys which makes the random function
very likely to select the same peers over and over.

This patch stores the keys reversed so the tag bytes are the least
significant ones. The peerbook is rewritten on startup if needed.