Check we have an inbound connection in stungun in the right place
Prior to the fix for tagging connection directions correctly, the
incorrect connection tracking allowed this code to usually work. However
we were checking for an inbound session in the wrong place. We really
care about dial confirmation from a peer we don't have an outbound
connection to. The prior check was checking the connection direction on
the reply from the intermediate host which, when you have no listen
addresses, will always be outbound, so stungun could never resolve the
NAT state.