Enable force setting network id
When a node isn't following the chain (like a seed node), then we
need a way to force set a network id so that the peerbook won't
store entries with an undefined network id, or a network id that
isn't the same as ours.

This leads to seed nodes holding entries and gossiping nodes
it shouldn't really be storing or gossiping.